Do not jit compile wrapped functions, run them via FLR instead. This is needed because certain ops that are placed on XLA devices do not support compilation. Running via FLR ensures we execute these ops the same way they would have run in eager mode. Long-term this should be handled by outside compilation but that is not available on GPU right now.
Place the NodeDef in the FunctionDef on the device that eager mode placement would have picked for the eager op. Certain ops e.g. RangeDataset are placed on GPU in eager mode but expect to be run on CPU. Eager mode placement understands this and picks the CPU device in this case, graph mode placer raises an error. For backwards compatibility we go with the choice made by eager mode placement. This requires us to trace a FunctionDef per-device. Optimizing the number of traces has been filed as future work.

PiperOrigin-RevId: 398228487
Change-Id: I9e9b4d7c0ccf8589f6f0bd312762eb978349fca7