Fix #51839: py_function records data for gradient when tape is not requested.
The memory leak test added for #35084 in tf.data still passes.

I renamed the confusing argument name eager= to use_eager_py_func=. The argument has little to do with eager mode; it only controls for which C++ op to use.

There may be a better way of fixing this. A few things to consider on top of
this CL:

- can we record the data actually to the tape instead of in our own cache?
  it feels like it is doable, but it is not so clear to me how to do this.

PiperOrigin-RevId: 398261944
Change-Id: I941bd9f7032ddfab209db161475c78068a17cc52