Move CollTaskParams into CollGroupParams
CollTaskParams only store whether a device is local or not. CollTaskParams::is_local corresponds to the devices, instead of the tasks. It makes more sense to fold CollTaskParams simply into CollGroupParams.

CollTaskParams used to populated as part of CompleteInstance. After the move it will be populated as par of CompleteGroup, which also makes more sense.

PiperOrigin-RevId: 398386333
Change-Id: Ic2f4f56bb8c9fecbc9e5edf0516c4f2648930659