Peel once on all tiled_loop dimensions
This CL assumes that the best single rule for loop peeling is to ensure that there is a single loop without any padding needed. All dimensions are peeled once to ensure that this loop exists. Other peeled loops could be peeled again to remove more padding, but the assumption is that in the common case, this will not be worth the IR size/compile time increase.

This is a temporary rule of thumb until more advanced heuristics can be created with multiple different code generation strategies depending on the input.

PiperOrigin-RevId: 399143335
Change-Id: Ie20653fc4d900c43b03107211c9cac4a3a89781c