Avoid crashing the process if FindClass or GetMethodID fail. After calls to FindClass or GetMethodID, check that there isn't an exception already pending before throwing an exception.
This ensures we propagate exceptions raised by FindClass or GetMethodID
back to the Java caller, rather than crashing the program by trying to
raise an exception when there is already an exception pending.

Also, add some additional error checking when looking up the XnnpackDelegate
class and constructor; this is more consistent with the code elsewhere, and
ensures that we don't construct and return a Delegate object when we're
throwing an exception, and also avoids trying to look up the constructor if we
can't find the class.

PiperOrigin-RevId: 399650055
Change-Id: I71f1e16b3ce18a165fdc0201825331033b11da9b