Turn `ConvolveExecutionPlan` into an interface for launching any DNN op.
This allows backends to hand over to clients ownership of opaque data needed to
launch the ops (currently convolution and fused convolution), without exposing
the details of that data in the backend-independent interface.  In particular
this is for cuDNN `ExecutionPlans`, but in principle it could be used for any
cached compiled kernels, allocated descriptor objects, etc.

PiperOrigin-RevId: 401320074
Change-Id: Ic2b24cf613a731d9d8976e27ddbdf3bcd33de905