Expose LookupGroup in CollectiveExecutor
This will be used by XLA to lookup collective group information. We can potentially also use it in collective ops once the refactoring of CompleteInstance is complete.

The signature is changed slightly. The original one is a little bit unclear which field of CollGroupParams should be set to call the method.

PiperOrigin-RevId: 404113861
Change-Id: Ia2fcfea4bac111cc42816ff26bd20509b95c3f6a