Disable test on the RUN line instead of excluding it in the BUILD file.
It is rare that we need to exclude a test in the build file (it is useful if it
hangs). Lit tests can be marked XFAIL / UNSUPPORTED instead.

Unfortunately this test only fails in assert mode, so it can't be marked "XFAIL".

// UNSUPPORTED: {{.*}}

is the syntax for lit, but google internal equivalent does not parse this, so we're
mangling the RUN line for now.

PiperOrigin-RevId: 404923736
Change-Id: Iee82476024f1f335b1ce769982773f5fb8674fa2