Update tf_generated_ops by running patch_tf_dialect.sh.
Note that with https://reviews.llvm.org/D111855, mlir-tblgen now expects that
strings are stored as their C++ representation. So instead of "foo", it now
expects the string "\"foo\"".
(This is analogous to how values are stored for other types. For example, we
  would store the string "42" for integers, not the (TableGen) integer 42.)
This makes it possible to store strings with special characters, and empty
strings. (The latter is needed for b/203010629.)

PiperOrigin-RevId: 406020184
Change-Id: Ia8c2d9aa317a97f9d1d26728609c4848e169efc1