Create a method to allow certain ops to be considered as fully-independent for automatic control dependencies, as a manual override. Fully-independent ops run in parallel with everything else: they are only guaranteed to terminate before the current function returns.
This is a manual override solution to be used while we sort out the side effect analysis rules around stateful ops such as WriteFile.

The method also marks the node with a special attribute that can be used in downstream analysis passes.

Note: this is not a public API, and is subject to change without notice.
PiperOrigin-RevId: 407224947
Change-Id: I1b0c55ecd01d186c8f8cbfffb6396889772ed6cb