Make functions subclass Trackable.
This paves the way to removing the special-casing for functions when saving and loading. In this CL, `list_functions` has been removed from the _AugmentedGraphView and combined with `list_children`, because with this change, functions are just another Trackable child.

Some unpleasantness when gathering the children: because functions were not previously trackable, classes that override `_list_functions_for_serialization` may be excluding tracked Functions. One example is Keras (the `train_function` is purposely excluded from `_list_functions_for_serialization`). These cases will need to be dealt with individually during the migrations.

PiperOrigin-RevId: 408723839
Change-Id: I46281899f8f49940671f80d2ac6ae77f348a38fe