Make functions subclass Trackable.
This paves the way to removing the special-casing for functions when saving and loading. In this CL, `list_functions` has been removed from the _AugmentedGraphView and combined with `list_children`, because with this change, functions are just another Trackable child.

Some unpleasantness when gathering the children: because functions were not previously trackable, classes that override `_list_functions_for_serialization` may be excluding tracked Functions. One example is Keras (the `train_function` is purposely excluded from `_list_functions_for_serialization`). These cases will need to be dealt with individually during the migrations.

Rollforward notes: Special case the "signatures" tf.function dependency. This follows the previous signatures validation behavior where `list_children` did not return tf.functions at all, so it did not error out if the root object had a function in the `signatures` attribute.

PiperOrigin-RevId: 409509058
Change-Id: Ia51bcb2b63fc348f4451ae53b006a2f7067a43ba