Fix a dead lock issue.
This dead lock is caused by deferred allocation. When there is no enough HBM for an allocation and there are pending deallocation, the allocation will block until more HBM is available after released pending deallocation. However, MakeTensorFromProto can hit this blocking allocation with a mutex, which cause the dead lock. Since the mutex is needed for releasing TPU buffer as well.

PiperOrigin-RevId: 410389327
Change-Id: I0e85337e996d8562e3d72be41602c5905711cc61