[XLA:CPU] Do not check that the size of the XLA parameter buffer is exactly equal to the size of the underlying given buffer
Instead, check that the underlying allocation is "large enough". This is also
more consistent with XLA:GPU behavior.

The mismatch can happen when the input comes from tf.where, which is backed by
an allocation larger than is actually required.

Providing tests is a bit of a chicken-and-an-egg problem: this is tested in
where ops tests, which are disabled due to this issue.

PiperOrigin-RevId: 411923802
Change-Id: Iee3da305b9692ff37bdecba058629419e3f495c3